requirements.txt
google-generativeai
mysql-connector-python
python-dotenv
flask
pandas
streamlit